+++ packages/example/android/app/src/main/kotlin/com/google/ml/kit/flutter/example/MainActivity.kt
@@ -2,4 +2,4 @@ package com.google.ml.kit.flutter.example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()